feat(Navigation): modularize and cleanup code - I91
